requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
timescaledb==0.2.0

# Redis and caching
redis==5.0.1
hiredis==2.2.3

# Task queue
celery==5.3.4
flower==2.0.1

# Data fetching and processing
yfinance==0.2.28
pandas==2.1.3
numpy==1.26.2
pandas-ta==0.3.14b0
ta-lib==0.4.28

# ML libraries (for future use)
tensorflow==2.15.0
scikit-learn==1.3.2

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
python-dateutil==2.8.2

# OpenAI (optional - for enhanced explanations)
openai==1.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
